WIP  - Implement Document Deletion on Frontend   - Implement Document Deletion on Frontend
