+++ gendox-frontend/src/utils/searchToolbar.js
@@ -1,0 +1,60 @@
+// ** MUI Imports
+import Box from '@mui/material/Box';
+import TextField from '@mui/material/TextField';
+import IconButton from '@mui/material/IconButton';
+import { GridToolbarFilterButton } from '@mui/x-data-grid';
+
+// ** Icon Imports
+import Icon from 'src/@core/components/icon';
+
+const quickSearchToolbar = (props) => {
+  return (
+    <Box
+      sx={{
+        gap: 2,
+        display: 'flex',
+        flexWrap: 'wrap',
+        alignItems: 'center',
+        justifyContent: 'space-between',
+        p: (theme) => theme.spacing(2, 5, 4, 5),
+      }}
+    >
+      {/* Filter Button */}
+      {/* <GridToolbarFilterButton /> */}
+      <Box>
+        
+      </Box>
+
+      {/* Search Input */}
+      <TextField
+        size="small"
+        value={props.value}
+        onChange={props.onChange}
+        placeholder="Search…"
+        InputProps={{
+          startAdornment: (
+            <Box sx={{ mr: 2, display: 'flex' }}>
+              <Icon icon="mdi:magnify" fontSize={20} />
+            </Box>
+          ),
+          endAdornment: (
+            <IconButton size="small" title="Clear" aria-label="Clear" onClick={props.clearSearch}>
+              <Icon icon="mdi:close" fontSize={20} />
+            </IconButton>
+          ),
+        }}
+        sx={{
+          width: {
+            xs: 1,
+            sm: 'auto',
+          },
+          '& .MuiInputBase-root > svg': {
+            mr: 2,
+          },
+        }}
+      />
+    </Box>
+  );
+};
+
+export default quickSearchToolbar;
